feat(i18n): create high order component for translations
